pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Slava Kovalchuk <Slava24122003@MAIL.RU>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
pydantic = "^2.9.2"
sqlalchemy = "^2.0.36"
asyncpg = "^0.30.0"
alembic = "^1.14.0"
loguru = "^0.7.2"
passlib = "^1.7.4"
bcrypt = "4.0.1"
python-multipart = "^0.0.17"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-dotenv = "^1.0.1"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
